PagedFile --> LinkPagedFile
Turned PagedFile into a ABC and my existing implementation into
LinkPagedFile. This opens the door for a simpler, IndexPagedFile that
will be more useful for efficiently dealing with LSM Tree sorted runs,
as these are static and won't require a lot of the overhead that the
linked paging brings with it.