Implemented unsorted memtable in LSM tree + cleanup
Added a constructor flag to use an unsorted memtable within the LSMTree
itself, and fixed a bug or two.

One of those bugfixes required tearing out the tombstone table stuff
temporarily. I'm going to replace this with a different data structure
anyway, so that was the most straightforward solution for now.