BugFix: Incorrect page pointers in ISAM Internal Nodes
Due to using the wrong variable, the first page pointer for each
internal node in the third level and above (aside from the first pointer
of the first node on each level) was inappropriately set. It pointed to
the target of its key within node it was supposed to point to, rather
than the node itself. This has now been rectified.

Additionally, the bypass code to circumvent this error is no longer
needed, and has been removed from LSMTree::is_deleted().