Intelligent Bloom Filter Allocation
The size of the bloom filter is now dynamically determined based on a
false positive rate and the number of tombstones on a given level.
Because of the specifics of the use of the filter, performing a lookup
against an empty filter will always return False--as an empty filter
means that there are no tombstones present.