ReadCache: Added a pin request counter
For counting the IO requests (total) required by the setup for variable
length keys, I've added a pin request counter to the read cache. This
counts every pin, regardless of whether it is a hit or miss in the
cache, to give a comparison metric for the worst case total IOs required
by the variable length key scheme.