Rejection Sampling for Unsorted Memtable
As an alternative to scanning the full memtable to construct a sample
range, rejection sampling can now be used. The full memtable (as of
sample range construction) is treated as the sampling range, and
standard rejection sampling is applied. This allows for much faster
sample range creation, at the cost of increased rejections during
sampling time.