DiskLevel: Corrected upper bound calculation
If the upper bound sampling key was larger than the largest key
within an ISAM Tree, the upper_bound routines would (correctly)
report that an upper bound didn't exist. But for a sample range,
this condition means that the upper bound should be the final page
within the tree.

Thus, to sample correctly, it was necessary to check for an INVALID_PNUM
result from upper_bound and, in that case, use the last leaf page as
the upper bound instead.