MemTableView: Added proper cutoffs
Now the "view" on the memtables will be properly fixed at creation time,
so no newly added records will be counted, messing up indexing and
exposing data that potentially shouldn't be in the sample.

The actual value of the cutoff is fixed just after each memtable is
pinned, and so it won't exactly reflect the state at the moment of pin,
but it will freeze the state to a particular moment in time, which
should be fine for our purposes.