lsm_mixed: added support for running against pre-existing tree
NOTE: This should work for the current case where there aren't
any tombstones in the pre-warmed up tree. But if the pre-warmed
up tree contains some tombstones, it'll break in two major ways.

First: there exists a chance for a record to be deleted twice (without
any reinsertion), because the process of selecting records to delete
doesn't know what records were pre-deleted during warmup.

Second: The record count of the tree is used to establish where in
the file to start reading new records from. Deletes will mess this up.

My proposed solution to both these problems is a slight change of
approach. Rather than loading x-percent of some file for warmup, and
then reading the remaining during the benchmark, what if we took our
data file and pre-split it into two. Then the warmup could process all
of the first file, and the benchmark can read the second. No need to
worry about the second issue.

For the first issue, we can simply use the range sample functionality
(assuming we know the full key range) to select a sample of records to
be staged for possible future deletion. The range sample procedure will
prevent us from staging an already deleted record for a second deletion,
and has the added benefit of automatically randomizing the list of
records, so the selection of records to delete process can be made more
efficient.