Throughput benches: adjusted delete batching
The original setup only pulled records to delete once, and so the deleted
records were only ever from the warmup data. Now it pulls records to
delete in smaller batches, more frequently, for a more even distribution
of deletes.

Additionally, even though the AVL/B+Tree varients do not need an exteral
tracking set to avoid duplicate deletes, I added this in for fairness,
as the time to manage this set was being counted against the LSM Tree
too. Given how bit the set potentially grew, this might have contributed
a non-negligible amount to the insertion latency for bigger data sets.